requirements.txt
Flask
Flask-SQLAlchemy
pywebview
pyinstaller
python-dotenv
Adafruit_PCA9685
qrcode[pil]
serial
requests
flask-socketio
Flask-JWT-Extended
gitpython
